Refactor to PoolStatusTest
